+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Sin publicar
+- Added optionals subtitle and button support to MLBusinessLoyaltyRingView.
+
 # v1.12.0
 🚀 1.12.0 🚀
 - Added getTouchpointViewHeight for Discount Touchpoints Component.
@@ -81,7 +84,7 @@
 # v1.0.9
 🚀 1.0.9 🚀
 - Change target version to iOS 10	- DownloadView improvements for small devices.
-- Fix an issue in MLBusinessDiscountBoxView in iOS 10	
+- Fix an issue in MLBusinessDiscountBoxView in iOS 10
 - L&F fixes
 
 # v1.0.8